+++ cmake/find_modules.cmake
@@ -18,7 +18,8 @@ set(CMAKE_MODULE_PATH "${ANAKIN_ROOT}/cmake")
 set(ANAKIN_LINKER_LIBS "")
 
 if(UNIX)
-	if(USE_ARM_PLACE)
+	if(USE_ARM_PLACE )
+	elseif(${CMAKE_SYSTEM_NAME} MATCHES "Darwin")
 	else()
 		find_library(RTLIB rt)
 		if(RTLIB)
