+++ saber/funcs/lstm.h
@@ -25,6 +25,7 @@
 
 #ifdef USE_X86_PLACE
 #include "saber/funcs/impl/x86/saber_lstm.h"
+#include "saber/funcs/impl/x86/vender_lstm.h"
 #endif
 
 #ifdef USE_ARM_PLACE
@@ -92,6 +93,7 @@ public:
                 this->_impl.push_back(new SaberLstm<TargetType, OpDtype, inDtype, outDtype, LayOutType_op, LayOutType_in, LayOutType_out>);
                 return SaberSuccess;
             case VENDER_IMPL:
+                LOG(INFO)<<"init vender";
                 this->_impl.push_back(new VenderLstm<TargetType, OpDtype, inDtype, outDtype, LayOutType_op, LayOutType_in, LayOutType_out>);
                 return SaberSuccess;
             default:
