+++ test/saber/x86/test_saber_func_lstm_x86.cpp
@@ -13,6 +13,7 @@
 #include "saber/saber_types.h"
 #include "x86_test_common.h"
 #include "test_saber_func_x86.h"
+#include "debug.h"
 
 using namespace anakin::saber;
 using namespace std;
@@ -295,7 +296,7 @@ bool lstm_test(test_lstm_params &param) {
     Tensor4f saberOutputh(outputShape);
     Tensor4f saberOutputc(outputShape);
     saber_outputs.push_back(&saberOutputh);
-    saber_outputs.push_back(&saberOutputc);
+//    saber_outputs.push_back(&saberOutputc);
 
     std::vector<Tensor4f*> ref_outputs;
     Tensor4f refOutputh(outputShape);
@@ -313,8 +314,64 @@ bool lstm_test(test_lstm_params &param) {
     saberLstm(inputs, saber_outputs, lstm_param, ctx_host);
 
     bool flag = compare_tensor(*saber_outputs[0], *ref_outputs[0], 1e-4);
-    flag &= compare_tensor(*saber_outputs[1], *ref_outputs[1], 1e-4);
+//    flag &= compare_tensor(*saber_outputs[1], *ref_outputs[1], 1e-4);
     return flag;
+}
+
+typedef Tensor<X86, AK_FLOAT, NCHW> TensorHf4;
+void py_lstm(int word_size = 222,
+             int hidden_size = 333){
+    Context<X86> ctx_dev(0, 1, 1);
+    std::vector<int> offsets = {0, 3,12,19,20};
+    ImplEnum test_mode=SABER_IMPL;
+//    ImplEnum test_mode=VENDER_IMPL;
+    bool is_reverse = false;
+    bool with_peephole=false;
+    Shape shape_weight(1, 1, 1,hidden_size*hidden_size*4+hidden_size*word_size*4);
+    Shape shape_bias;
+    if(with_peephole){
+        shape_bias=Shape(1,1,1,hidden_size*7);
+    }else{
+        shape_bias=Shape(1,1,1,hidden_size*4);
+    }
+    Shape shape_x(offsets[offsets.size() - 1], word_size, 1, 1);
+    Shape shape_h(offsets[offsets.size() - 1], hidden_size, 1, 1);
+    TensorHf4 host_x(shape_x);
+    TensorHf4 host_weight(shape_weight);
+    TensorHf4 host_bias(shape_bias);
+    TensorHf4 host_hidden_out(shape_h);
+    readTensorData(host_weight, "host_w");
+    readTensorData(host_x, "host_x");
+    readTensorData(host_bias, "host_b");
+
+    host_x.set_seq_offset(offsets);
+    LstmParam<TensorHf4> param(&host_weight, &host_bias,nullptr,Active_unknow,Active_sigmoid,Active_tanh,Active_tanh,
+                               with_peephole,false,is_reverse);
+    Lstm<X86, AK_FLOAT, AK_FLOAT, AK_FLOAT, NCHW, NCHW, NCHW> lstm_op;
+
+    std::vector<TensorHf4*> inputs;
+    std::vector<TensorHf4*> outputs;
+    inputs.push_back(&host_x);
+    outputs.push_back(&host_hidden_out);
+
+    SABER_CHECK(lstm_op.init(inputs, outputs, param, SPECIFY, test_mode, ctx_dev));
+    SABER_CHECK(lstm_op.compute_output_shape(inputs, outputs, param));
+    outputs[0]->re_alloc(outputs[0]->valid_shape());
+    SABER_CHECK(lstm_op(inputs, outputs, param, ctx_dev));
+
+    TensorHf4 compare_g(shape_h);
+    readTensorData(compare_g, "host_correct");
+    write_tensorfile(host_hidden_out, "host_g.txt");
+    write_tensorfile(compare_g, "host_correct.txt");
+    double maxdiff = 0;
+    double maxratio = 0;
+    tensor_cmp_host(host_hidden_out.data(), compare_g.data(), host_hidden_out.valid_size(), maxratio, maxdiff);
+    if (abs(maxratio) <= 0.001) {
+        LOG(INFO) << "passed  " << maxratio<<","<<maxdiff<<",?="<<abs(maxratio);
+    } else {
+        LOG(INFO) << "failed : ratio " << maxratio<<","<<maxdiff;
+    }
+
 }
 
 TEST(TestSaberFuncX86, test_tensor_lstm) {
@@ -324,20 +381,20 @@ TEST(TestSaberFuncX86, test_tensor_lstm) {
         // batch_size, input_size, layer_size, input_activation, gate_activation, candidate_activation, cell_activation, with_peephole, with_init_hidden, skip_input
         test_lstm_params{6, 55, 300, Active_unknow, Active_sigmoid, Active_relu, Active_sigmoid, false, false, false},
         test_lstm_params{6, 55, 300, Active_unknow, Active_sigmoid, Active_relu, Active_sigmoid, true, false, false},
-        test_lstm_params{6, 55, 300, Active_unknow, Active_sigmoid, Active_relu, Active_sigmoid, false, true, false},
-        test_lstm_params{6, 55, 300, Active_unknow, Active_sigmoid, Active_relu, Active_sigmoid, true, true, false},
-        test_lstm_params{6, 55, 300, Active_tanh, Active_sigmoid, Active_relu, Active_sigmoid, false, false, false},
-        test_lstm_params{6, 55, 300, Active_tanh, Active_sigmoid, Active_relu, Active_sigmoid, true, false, false},
-        test_lstm_params{6, 55, 300, Active_tanh, Active_sigmoid, Active_relu, Active_sigmoid, true, true, false},
-        test_lstm_params{6, 55, 300, Active_tanh, Active_sigmoid, Active_relu, Active_sigmoid, false, true, false},
-        test_lstm_params{6, 55, 300, Active_stanh, Active_sigmoid, Active_relu, Active_sigmoid, false, false, false},
-        test_lstm_params{6, 55, 300, Active_stanh, Active_sigmoid, Active_relu, Active_sigmoid, true, false, false},
-        test_lstm_params{6, 55, 300, Active_stanh, Active_sigmoid, Active_relu, Active_sigmoid, false, true, false},
-        test_lstm_params{6, 55, 300, Active_stanh, Active_sigmoid, Active_relu, Active_sigmoid, true, true, false},
-        test_lstm_params{6, 55, 300, Active_unknow, Active_sigmoid, Active_relu, Active_sigmoid, false, false, true},
-        test_lstm_params{6, 55, 300, Active_unknow, Active_sigmoid, Active_relu, Active_sigmoid, true, false, true},
-        test_lstm_params{6, 55, 300, Active_unknow, Active_sigmoid, Active_relu, Active_sigmoid, false, true, true},
-        test_lstm_params{6, 55, 300, Active_unknow, Active_sigmoid, Active_relu, Active_sigmoid, true, true, true},
+//        test_lstm_params{6, 55, 300, Active_unknow, Active_sigmoid, Active_relu, Active_sigmoid, false, true, false},
+//        test_lstm_params{6, 55, 300, Active_unknow, Active_sigmoid, Active_relu, Active_sigmoid, true, true, false},
+//        test_lstm_params{6, 55, 300, Active_tanh, Active_sigmoid, Active_relu, Active_sigmoid, false, false, false},
+//        test_lstm_params{6, 55, 300, Active_tanh, Active_sigmoid, Active_relu, Active_sigmoid, true, false, false},
+//        test_lstm_params{6, 55, 300, Active_tanh, Active_sigmoid, Active_relu, Active_sigmoid, true, true, false},
+//        test_lstm_params{6, 55, 300, Active_tanh, Active_sigmoid, Active_relu, Active_sigmoid, false, true, false},
+//        test_lstm_params{6, 55, 300, Active_stanh, Active_sigmoid, Active_relu, Active_sigmoid, false, false, false},
+//        test_lstm_params{6, 55, 300, Active_stanh, Active_sigmoid, Active_relu, Active_sigmoid, true, false, false},
+//        test_lstm_params{6, 55, 300, Active_stanh, Active_sigmoid, Active_relu, Active_sigmoid, false, true, false},
+//        test_lstm_params{6, 55, 300, Active_stanh, Active_sigmoid, Active_relu, Active_sigmoid, true, true, false},
+//        test_lstm_params{6, 55, 300, Active_unknow, Active_sigmoid, Active_relu, Active_sigmoid, false, false, true},
+//        test_lstm_params{6, 55, 300, Active_unknow, Active_sigmoid, Active_relu, Active_sigmoid, true, false, true},
+//        test_lstm_params{6, 55, 300, Active_unknow, Active_sigmoid, Active_relu, Active_sigmoid, false, true, true},
+//        test_lstm_params{6, 55, 300, Active_unknow, Active_sigmoid, Active_relu, Active_sigmoid, true, true, true},
         /*test_lstm_params{6, 55, 300, Active_tanh, Active_sigmoid, Active_relu, Active_sigmoid, false, false, true},
         test_lstm_params{6, 55, 300, Active_tanh, Active_sigmoid, Active_relu, Active_sigmoid, true, false, true},
         test_lstm_params{6, 55, 300, Active_tanh, Active_sigmoid, Active_relu, Active_sigmoid, true, true, true},
@@ -358,6 +415,8 @@ TEST(TestSaberFuncX86, test_tensor_lstm) {
             LOG(ERROR) << "Test Failed";
         }
     }
+
+//    py_lstm();
 }
 
 int main(int argc, const char** argv) {
